+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## UNRELEASED
+- Added Courier Death Notice. 
 - Reworked how delivery route and delivery duration is calculated. 
   - It's somewhat involved, maybe I'll explain it the future in a wiki or something. 
 - Reduced `courier_travel_speed` config value from 25 to 20.